requirements.txt
google-assistant-grpc==0.1.0
google-cloud-speech==0.30.0
google-auth-oauthlib==0.2.0
pyasn1==0.4.2
grpcio==1.7.0
google-api-python-client==1.7.3
oauth2client==4.1.2
gspread==3.0.1
absl-py==0.2.2